Use route vars over JSON values for parsing the ID.
The only handler that needs an identifier in the JSON is Add().
The rest of the handlers get from the identifier route variable
that is parsed from the URL. As a result, the Get() and Delete()
handlers no longer require JSON to be passed to them.

Delete() was fixed so that calls will succeed even if nothing is
done.